refactor: to only use one slug
